Move tests to a common directory
