+++ .gitignore
@@ -32,16 +32,16 @@ include
 *.i*86
 *.x86_64
 *.hex
-/example_kem
-/speed_kem
-/test_kem
-/test_kex
-/test_sig
-/test_rand
-/test_aes
-/test_sha3
-/minimal_kex_oqs
-/minimal_sig_oqs
+tests/example_kem
+tests/speed_kem
+tests/test_kem
+tests/test_kex
+tests/test_sig
+tests/test_rand
+tests/test_aes
+tests/test_sha3
+tests/minimal_kex_oqs
+tests/minimal_sig_oqs
 
 # Debug files
 *.dSYM/
