+++ tests/test_rand.c
